Container File Support
This commit adds the file block to the lxd_container resource.
One or more individual files may be defined in the resource and
they will be uploaded/pushed to the container via the LXC/LXD API.

Updates to the files will cause all old files to be deleted and
new copies to be uploaded.