Acc Tests: Disable second network in access interface test
Due to an unidentified timing issue, sometimes an LXD network
is not being created in time for the LXD container to launch
and is causing an out-of-ordering error to happen with the tests.

For now, we'll remove the second network from the test so there's
only one network to compare for test output.