Scan for shortest IPv6 address
This change returnes the shortest IPv6 address for a container instead
of the first one. This makes the order in which LXD returns IP address
less important. IP addresses are compared by the length of their string
representation. Manually assigned static IPv6 addresses, that often are
shorter then automatically generated addresses, are preferably returned
this way.