Expose containers IPv6 address (#173)
* Expose containers IPv6 address in resource attributes

This commit adds the IPv6 address to the container resource attributes.
It looks on the same interface as for the IPv4 address and takes the
first IPv6 address.

The test uses a fixed hwaddr to always have the same IPv6 address
assigned via SLAAC.

* Scan for shortest IPv6 address

This change returnes the shortest IPv6 address for a container instead
of the first one. This makes the order in which LXD returns IP address
less important. IP addresses are compared by the length of their string
representation. Manually assigned static IPv6 addresses, that often are
shorter then automatically generated addresses, are preferably returned
this way.

* Update documentation for IPv6 addresses

* Select IPv6 as IPv4 but favor global addresses

This changes the IPv6 address selection algorithm to work like for IPv4.
The only difference is that global addresses are favored over link-local
addresses. The latter are only selected when no global address is
present.