Allow updating networks in-place (#210)
* Allow updating networks in-place

This commit adds an update method to the network resource that applies
config changes without recreating the network. This allows updating
networks that are in use.

* Add tests for updating network in-place

This test is a bit special as it needs an additional container. Without
the container using the network, the test would pass without the
in-place update code too, because Terraform would just delete the old
network and create new one with an identical name.

With an attached container, trying to delete the network would fail.
This way the test ensures that the network is actually updated in-place.