Ensure storage pool tests use an empty directory (#218)
* Ensure storage pool tests use an empty directory

All tests involving a storage pool used `/mnt` has the hard-coded source
for the pool. This fails if the hosts `/mnt` is *not* empty. This might
actually not be true and in some cases (e.g. Github Actions VM) it
cannot be changed at all by the developer.

This CL changes all tests to use a new temporary directory, which
*always* is empty as the pool source.

The Go 1.15 update is needed as with 1.15 the testing frameworks
provides a ready to use function to get a temporary directory in a test
case.

* Add workflow to run tests using Github Actions

This commit adds a workflow to run all provider tests using the Github
Actions pipeline.