Fix file upload on virtual machine creation
Creating a new instance that has `file {` blocks doesn't work if the
type is "virtual-machine" because it relies on LXD Agent running.

Modify resourceLxdInstanceCreate to ensure the LXD Agent is running
before trying to upload files. All other cases remain unmodified, and
the user is responsible for ensuring the VM is running when making
changes.