internal/instance: Export only global IP addresses from a single interface
Previously the IPv4 and IPv6 addresses could be returned from different interfaces. If IPv6 was
not found for a particular interface it would be searched further on other interfaces. Now, the
IPs are returned either from an interface defined in user.access_interface or from the first
interface that contains a global address (sorted alphabetically).

Signed-off-by: Din Music <din.music@canonical.com>