instance: Tweak default interface logic
This reworks the logic used to determine the default ipv4, ipv6 and mac
address to return.

Rather than picking the last values for each type, this now sorts the
whole list of interface from most to least desirable and then selects
the most desirable interface from which all the information is
retrieved.

It also always restricts itself to global addresses.

Closes #73

Signed-off-by: Stéphane Graber <stgraber@stgraber.org>