internal/instance: Handle cluster group targets and in-cluster migrations
Target attribute now correctly handels cluster groups (e.g. @my-group).

A new attribute "allow_restart" is also introduced and is used to give the provider
a permission to temporarily stop the instance when doing instance renaming or migration.
This only applies when the instance is currently running and is not planned to be stopped.

Signed-off-by: Din Music <din.music@canonical.com>